# api/openapi.yaml
openapi: 3.0.3
info:
  title: Assay API
  description: LLM 渠道检测平台 API。本文件是前后端契约的唯一真相源（契约优先）。
  version: 0.2.0
servers:
  - url: /api
paths:
...
  /users:
    get:
      operationId: listUsers
      summary: 用户列表（需 users 权限）
      tags: [users]
      responses:
        "200":
          description: 全部用户
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createUser
      summary: 创建用户（需 users 权限；不开放注册，账号只能在此创建）
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: 用户名已存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{id}:
    patch:
      operationId: updateUser
      summary: 修改用户角色或重置密码（需 users 权限；不能改自己的角色）
      tags: [users]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: 修改后的用户
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteUser
      summary: 删除用户（需 users 权限；不能删除自己）
      tags: [users]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: 删除成功
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: 请求非法
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: 未登录或会话已过期
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: 无权访问该模块
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    User:
      type: object
      required: [id, username, roleId, roleName, createdAt]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        roleId:
          type: string
          format: uuid
        roleName:
          type: string
        createdAt:
          type: string
          format: date-time
    UserCreate:
      type: object
      required: [username, password, roleId]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64
        password:
          type: string
          minLength: 8
        roleId:
          type: string
          format: uuid
    UserUpdate:
      type: object
      properties:
        roleId:
          type: string
          format: uuid
        password:
          type: string
          minLength: 8
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string